requirements.txt
Flask>=3.0  
langchain>=0.2    
chromadb>=0.5    
sentence-transformers>=2.7 
pypdf>=4.0       
openai>=1.30        
tiktoken>=0.6      
python-dotenv>=1.0  
langchain-community
rank_bm25
numpy<2.0   
langchain-huggingface
flask-cors>=4.0        
langchain-openai